requirements.txt
streamlit>=1.30
pandas>=2.1
openpyxl>=3.1
pyyaml>=6.0
pytest>=7.4
